Rest controller methods were added

+++ src/main/java/com/sergosoft/goodscatalog/service/impl/ProductServiceImpl.java
@@ -2,8 +2,14 @@ package com.sergosoft.goodscatalog.service.impl;
 
 import java.util.List;
 
+import com.sergosoft.goodscatalog.dto.product.ProductDto;
+import com.sergosoft.goodscatalog.dto.product.ProductFilter;
+import com.sergosoft.goodscatalog.service.specification.ProductSpecification;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import com.sergosoft.goodscatalog.dto.product.ProductRequest;
@@ -22,8 +28,20 @@ public class ProductServiceImpl implements ProductService {
     private final ProductMapper productMapper;
 
     @Override
+    public Page<ProductDto> getFilteredProductsByPage(int page, int pageSize, ProductFilter filter) {
+        log.debug("Fetching filtered products on page {} (page size {})", page, pageSize);
+        Specification<Product> specification = Specification
+                .where(ProductSpecification.hasPriceBetween(filter.getMinPrice(), filter.getMaxPrice()));
+
+        Page<Product> productsPage = productRepository.findAll(specification, PageRequest.of(page, pageSize));
+        Page<ProductDto> productDtoPage = productsPage.map(productMapper::toDto);
+        log.info("Retrieved {} products", productDtoPage.getSize());
+        return productDtoPage;
+    }
+
+    @Override
     public List<Product> getAllProducts() {
-        log.debug("Fetching all products");
+        log.debug("Fetching all products...");
         List<Product> products = productRepository.findAll();
         log.info("Retrieved {} products", products.size());
         return products;
@@ -47,23 +65,28 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void updateProduct(Long id, ProductRequest productRequest) {
+    public ProductDto updateProduct(Long id, ProductRequest productRequest) {
         log.debug("Updating product with ID: {}", id);
         ensureProductExists(id);
 
         Product product = productMapper.toEntity(productRequest);
         product.setId(id);
 
-        productRepository.save(product);
+        Product savedProduct = productRepository.save(product);
         log.info("Updated product with ID: {}", id);
+        return productMapper.toDto(savedProduct);
     }
 
     @Override
-    public void deleteProduct(Long id) {
+    public boolean deleteProduct(Long id) {
         log.debug("Deleting product with ID: {}", id);
-        ensureProductExists(id);
+        if(!productRepository.existsById(id)) {
+            log.debug("Unable to delete nonexistent product(id={})", id);
+            return false;
+        }
         productRepository.deleteById(id);
         log.info("Deleted product with ID: {}", id);
+        return true;
     }
 
     private Product findProductOrThrow(Long id) {
